requirements.txt
langchain-mcp-adapters
python-dotenv
pyairtable
e2b-code-interpreter